图片这一块，接入了mingdfs，并且已经联调成功，整体上没什么太大的问题，然后，删除栏目时，需要再接入mingmq，因为删除栏目，那么栏目下的所有图片都要被删除掉，这样才是合理的，但是，我总不能在一个http请求里删除所有文件吧？因为这个删除文件首先是要将该栏目的所有文件从数据库中查询出来，然后，再向mingdfs发送请求删除，由于文件如果有几千万，显然依靠批量删除是不现实的，所以，这时候必须启动一个消息队列，让reborn将要删除的文件，推向消息队列，这样，让消息队列去删那成千上万的文件，这样是比较合理的。
图片这一块，也就是这个删除栏目没做好了，所以，接入消息队列正好是当下应该做的事。

首先在reborn上写好mingmq的配置，然后，在reborn __init__里面初始化这个消息队列连接池，再启动一个消息队列，启动消息队列之后，用户发起http请求删除栏目时，我首先将该栏目的id查询出来，然后再将该category_id推向消息队列，消息队列拿到这个任务数据之后，分页查询出该category_id的所有图片行，然后，向mingdfs发送http delete文件请求。

其实我觉得有必要设计两个消息队列，第一个消息队列是用于分发请求，也就是拿到category_id之后，从数据库中查询出图片行，然后将这些图片行分发到用于请求mingdfs delete的队列，这样，请求mingdfs 队列失败之后，就不用确认，成功就确认该行，一段时间之后，mingdfs自己就将这些任务重新推入队列，这样就基本能确保数据的一致性。如果还是不行，就人工去删数据库就可以了。

我想这就是这个世界还需要程序员的原因。

所以，具体的做事方法已经确定好了：
第一、消息队列的配置在reborn里面，一个是reborn用于推送任务的连接池，另外两个是reborn里单独进程，也就是分发删除任务和用于处理删除任务的队列配置。
第二、写好了这些配置，我们还要接着写reborn里的推送任务的代码，两个消息队列的代码。
第三、写好了reborn项目中的东西之后，我觉得有必要给mingmq设置一个任务重新推送到队列的一个时间间隔，所以，还需要给mingmq设置一个配置，然后，在mingmq admin那里能看到这个配置，并且能修改，并且，mingmq能全部重新推送未确认的任务重新推送到队列。

所以，今天的事情还是很多的。

+++ reborn/apps/image/photo.py
@@ -1,14 +1,18 @@
 import os
 import time
+import traceback
 
-from flask import request, session, Blueprint, send_from_directory, redirect, url_for
+from flask import request, session, Blueprint, redirect, url_for
 from werkzeug.utils import secure_filename
 
 from reborn import ACCOUNT_MYSQL_POOL, IMAGE_MYSQL_POOL
 from reborn.db.account import User
 from reborn.db.image import Photo, Category
+from reborn.settings.apps import CACHE_DIR, MDFS_API_KEY, MDFS_DOWNLOAD_URL, MDFS_EDIT_URL, MDFS_UPLOAD_URL, \
+    MDFS_DOWNLOAD_MANY_URL, MDFS_DELETE_URL
 from reborn.settings.apps.account import IS_LOGIN
-from reborn.settings.apps.image import UPLOAD_FOLDER
+from reborn.utils.mdfs import download as mdfs_download, upload as mdfs_upload, download_many as mdfs_download_many, \
+    edit as mdfs_edit, delete as mdfs_delete
 
 IMAGE_PHOTO_BP = Blueprint('image_photo_bp', __name__)
 
@@ -35,28 +39,20 @@ def del_photo():
             photo = Photo(IMAGE_MYSQL_POOL)
             result = photo.get_photo(title, user_id)
             if result:
+                third_user_id = user_id
+                category_id = result[0]['category_id']
+                title = result[0]['title']
+                date = result[0]['date']
+                category_name = result[0]['category_name']
                 try:
-                    local_url = result[0]
-                    filename = UPLOAD_FOLDER + os.path.sep + local_url['local_url']
-                    os.remove(filename)
-                    print('删除文件成功：', filename)
-
+                    if 0 == mdfs_delete(MDFS_DELETE_URL, MDFS_API_KEY, third_user_id, category_id, title):
+                        return {"data": [], "status": -1}
                     result = photo.delete_photo(title, user_id)
-
-                    return {
-                        "data": [],
-                        "status": result
-                    }
-                except OSError as err:
-                    print(err)
-                    print('删除文件失败：', filename)
-            else:
-                return {
-                    "data": [],
-                    "status": -1
-                }
-        else:
-            return redirect(url_for("user_bp.login"))
+                    return {"data": [], "status": result}
+                except:
+                    print(traceback.format_exc())
+            return {"data": [], "status": -1}
+        return redirect(url_for("user_bp.login"))
 
 
 @IMAGE_PHOTO_BP.route('/get_photo', methods=['POST', 'GET'])
@@ -70,8 +66,7 @@ def get_photo():
                                 "title": xxx,
                                 "date": xxx,
                                 "category_name": xxx,
-                                "local_url": xxx,
-                                "remote_url": xxx
+                                "url": xxx,
                             }],
                             "status": 1
                         }
@@ -84,8 +79,7 @@ def get_photo():
                                 "title": xxx,
                                 "date": xxx,
                                 "category_name": xxx,
-                                "local_url": xxx,
-                                "remote_url" xxx
+                                "url": xxx,
                             }],
                             "status": 1
                         }
@@ -99,18 +93,21 @@ def get_photo():
             photo = Photo(IMAGE_MYSQL_POOL)
             result = photo.get_photo(title, user_id)
             if result:
-                return {
-                    "data": result,
-                    "status": 1
-                }
-            else:
-                return {
-                    "data": [],
-                    "status": -1
-                }
-        else:
-            return redirect(url_for("user_bp.login"))
+                third_user_id = user_id
+                category_id = result[0]['category_id']
+                title = result[0]['title']
+                date = result[0]['date']
+                category_name = result[0]['category_name']
+                url = None
+                try:
+                    url = mdfs_download(MDFS_DOWNLOAD_URL, MDFS_API_KEY, third_user_id, category_id, title)
+                except:
+                    pass
 
+                if url is None: return {"data": [], "status": -1}
+                result = [{'title': title, 'date': date, 'category_name': category_name, 'url': url}]
+                return {"data": result, "status": 1}
+            return {"data": [], "status": -1}
     elif request.method == "GET":
         look = request.args.get("look")
         if look and look.strip() != "":
@@ -122,25 +119,23 @@ def get_photo():
                 photo = Photo(IMAGE_MYSQL_POOL)
                 result = photo.get_photo(title, vi_user_id)
                 if result is not None:
-                    return {
-                        "data": result,
-                        "status": 1
-                    }
-                else:
-                    return {
-                        "data": [],
-                        "status": -1
-                    }
-            else:
-                return {
-                    "data": [],
-                    "status": -1
-                }
-        else:
-            return {
-                "data": [],
-                "status": -1
-            }
+                    third_user_id = vi_user_id
+                    category_id = result[0]['category_id']
+                    title = result[0]['title']
+                    date = result[0]['date']
+                    category_name = result[0]['category_name']
+                    url = None
+                    try:
+                        url = mdfs_download(MDFS_DOWNLOAD_URL, MDFS_API_KEY, third_user_id, category_id, title)
+                    except:
+                        pass
+                    if url is None:
+                        return {"data": [], "status": -1}
+                    result = [{'title': title, 'date': date, 'category_name': category_name, 'url': url}]
+                    return {"data": result, "status": 1}
+                return {"data": [], "status": -1}
+
+    return redirect(url_for("user_bp.login"))
 
 
 @IMAGE_PHOTO_BP.route('/pag_photo', methods=['POST', 'GET'])
@@ -151,14 +146,14 @@ def pag_photo():
     GET 获取未登录用户的图片列表
         url参数 ?look=xxx
         请求form表单 {"page": 1, "category_name": xxx}
-        返回json 成功: {"data": [{"title": xxx, "local_url": xxx, "remote_url": xxx, "date": xxx}],
+        返回json 成功: {"data": [{"title": xxx, "url": xxx, "date": xxx}],
                                 "status": 1}
                 失败: {"data": [], "status": -1}
 
 
     POST 获取已登录用户的图片列表
         请求form表单 {"page": 1, "category_name": xxx}
-        返回json 成功: {"data": [{"title": xxx, "local_url": xxx, "remote_url": xxx, "date": xxx}]
+        返回json 成功: {"data": [{"title": xxx, "date": xxx, 'url': xxx, 'category_name': xxx}]
                                 "status": 1}
                 失败: {"data": [], "status": -1}
     """
@@ -171,20 +166,33 @@ def pag_photo():
             photo = Photo(IMAGE_MYSQL_POOL)
             result = photo.pag_photo(page, category_name, user_id)
             if result is not None:
-                return {
-                    "data": result,
-                    "status": 1
-                }
-            else:
-                return {
-                    "data": [],
-                    "status": -1
-                }
-        else:
-            return {
-                "data": [],
-                "status": -1
-            }
+                req_data = []
+                for row in result:
+                    title = row['title']
+                    category_id = row['category_id']
+                    req_data.append({
+                        'third_user_id': user_id,
+                        'category_id': category_id,
+                        'title': title,
+                    })
+                res_data = mdfs_download_many(MDFS_DOWNLOAD_MANY_URL, MDFS_API_KEY, req_data)
+                if res_data is None:
+                    return {"data": [], "status": -1}
+
+                response_data = []
+                for row in result:
+                    for res_d in res_data:
+                        if row['title'] == res_d['title'] and str(row['category_id']) == str(res_d['category_id']) \
+                                and str(user_id) == str(res_d['third_user_id']):
+                            response_data.append({
+                                'category_name': row['category_name'],
+                                'url': res_d.get('url', None),
+                                'title': row['title'],
+                                'date': row['date']
+                            })
+                            break
+                return {"data": response_data, "status": 1}
+        return {"data": [], "status": -1}
 
     elif request.method == "GET":
         look = request.args.get("look")
@@ -198,25 +206,33 @@ def pag_photo():
                 result = photo.pag_photo(page, category_name, vi_user_id)
 
                 if result is not None:
-                    return {
-                        "data": result,
-                        "status": 1
-                    }
-                else:
-                    return {
-                        "data": [],
-                        "status": -1
-                    }
-            else:
-                return {
-                    "data": [],
-                    "status": -1
-                }
-        else:
-            return {
-                "data": [],
-                "status": -1
-            }
+                    req_data = []
+                    for row in result:
+                        title = row['title']
+                        category_id = row['category_id']
+                        req_data.append({
+                            'third_user_id': vi_user_id,
+                            'category_id': category_id,
+                            'title': title,
+                        })
+                    res_data = mdfs_download_many(MDFS_DOWNLOAD_MANY_URL, MDFS_API_KEY, req_data)
+                    if res_data is None:
+                        return {"data": [], "status": -1}
+
+                    response_data = []
+                    for row in result:
+                        for res_d in res_data:
+                            if row['title'] == res_d['title'] and str(row['category_id']) == str(res_d['category_id']) \
+                                    and str(vi_user_id) == str(res_d['third_user_id']):
+                                response_data.append({
+                                    'category_name': row['category_name'],
+                                    'url': res_d['url'],
+                                    'title': row['title'],
+                                    'date': row['date']
+                                })
+                                break
+                    return {"data": response_data, "status": 1}
+        return {"data": [], "status": -1}
 
 
 @IMAGE_PHOTO_BP.route('/upload', methods=['POST'])
@@ -238,64 +254,46 @@ def upload():
             photo = Photo(IMAGE_MYSQL_POOL)
 
             if 1 == photo.exist(title, user_id):
-                return {
-                    "data": [],
-                    "status": -1
-                }
+                return {"data": [], "status": -1}
 
             category = Category(IMAGE_MYSQL_POOL)
 
-            if -1 == category.exist(category_name, user_id):
-                return {
-                    "data": [],
-                    "status": -1
-                }
+            category_id = category.get_category_id(category_name, user_id)
+            if category_id is None:
+                return {"data": [], "status": -1}
 
             if 'upload_file_name' not in request.files:
-                return {
-                    "data": [],
-                    "status": -1
-                }
+                return {"data": [], "status": -1}
 
             file = request.files['upload_file_name']
-
             if file.filename == '':
-                return {
-                    "data": [],
-                    "status": -1
-                }
-            filename = str(time.time()) + '_' + secure_filename(file.filename)
-            if file and _allowed_file(file.filename):
-                file.save(os.path.join(UPLOAD_FOLDER, filename))
-            else:
-                return {
-                    "data": [],
-                    "status": -1
-                }
+                return {"data": [], "status": -1}
 
-            if 1 == photo.add_photo(title, category_name, filename, user_id):
-                return {
-                    "data": [],
-                    "status": 1
-                }
-            return {
-                "data": [],
-                "status": -1
-            }
+            filename = str(time.time()) + '_' + secure_filename(file.filename)
+            file_extension = ''
+            fn_list = filename.rsplit('.', 1)
+            if len(fn_list) > 1:
+                file_extension = fn_list[1]
+
+            if file and _allowed_file(file.filename):
+                file_path = os.path.join(CACHE_DIR, filename)
+                file.save(file_path)
+                status = mdfs_upload(MDFS_UPLOAD_URL, MDFS_API_KEY, user_id, category_id, title, filename, file_path)
+                if status == 0:
+                    return {"data": [], "status": -1}
+            else:
+                return {"data": [], "status": -1}
+
+            if 1 == photo.add_photo(title, category_name, file_extension, user_id):
+                return {"data": [], "status": 1}
+            return {"data": [], "status": -1}
         else:
             return redirect(url_for("user_bp.login"))
 
 
-@IMAGE_PHOTO_BP.route('/view_photo/<filename>', methods=['GET'])
-def view_photo(filename):
-    if request.method == 'GET':
-        return send_from_directory(UPLOAD_FOLDER, filename)
-
-
 @IMAGE_PHOTO_BP.route('/update_photo', methods=['POST'])
 def update_photo():
-    """
-    更新图片
+    """更新图片
 
     POST 请求form表单 {"src_title": xxx, "new_title": xxx, "category_name": xxx, "upload_file_name": xxx}
          返回json 成功: {"data": [], "status": 1}
@@ -311,54 +309,77 @@ def update_photo():
 
             photo = Photo(IMAGE_MYSQL_POOL)
 
-            if 1 == photo.exist(new_title, user_id) and src_title != new_title or -1 == photo.exist(src_title, user_id):
-                return {
-                    "data": [],
-                    "status": -1
-                }
+            src_photo_id = photo.get_photo_id(src_title, user_id)
+            new_photo_id = photo.get_photo_id(new_title, user_id)
+            if src_photo_id != new_photo_id:
+                if src_photo_id is None:
+                    return {"data": [], "status": -1}
 
             category = Category(IMAGE_MYSQL_POOL)
-
-            if -1 == category.exist(category_name, user_id):
-                return {
-                    "data": [],
-                    "status": -1
-                }
+            category_id = category.get_category_id(category_name, user_id)
+            if category_id is None:
+                return {"data": [], "status": -1}
 
             if 'upload_file_name' not in request.files:
-                return {
-                    "data": [],
-                    "status": -1
-                }
-
+                return {"data": [], "status": -1}
             file = request.files['upload_file_name']
+
+            src_img = photo.get_photo(src_title, user_id)
+            if src_img is None:
+                return {"data": [], "status": 0}
 
             if file.filename == '':
                 # 更新图片标题，时间，分类
-                result = photo.update_photo(user_id, src_title, new_title, category_name)
-                return {
-                    "data": [],
-                    "status": result
-                }
-            else:
-                # 删除老图片，更新标题，时间，分类，local_url，保存新图片
-                img = photo.get_photo(src_title, user_id)
-                local_url = img[0]['local_url']
-                os.remove(UPLOAD_FOLDER + os.path.sep + local_url)
+                src_third_user_id = user_id
+                src_category_id = src_img[0]['category_id']
+                src_file_extension = src_img[0]['file_extension']
+                new_third_user_id = user_id
+                new_category_id = category_id
+                new_file_extension = src_file_extension
+                try:
+                    if 0 == mdfs_edit(MDFS_EDIT_URL, MDFS_API_KEY, src_third_user_id, src_category_id, src_title,
+                                      src_file_extension,
+                                      new_third_user_id, new_category_id, new_title, new_file_extension,
+                                      upload_file_name=None, upload_file_path=None):
+                        return {"data": [], "status": -1}
+                except:
+                    print('文件在mdfs上修改失败')
+                    return {"data": [], "status": -1}
 
-                # 保存新图片
+                result = photo.update_photo(user_id, src_title, new_title, category_name)
+
+                return {"data": [], "status": result}
+            else:
                 filename = str(time.time()) + '_' + secure_filename(file.filename)
+                new_file_extension = ''
+                fn_list = filename.rsplit('.', 1)
+                if len(fn_list) == 2:
+                    new_file_extension = fn_list[1]
+
+                file_path = os.path.join(CACHE_DIR, filename)
                 if file and _allowed_file(file.filename):
-                    file.save(os.path.join(UPLOAD_FOLDER, filename))
-                    result = photo.update_photo(user_id, src_title, new_title, category_name, filename)
-                    return {
-                        "data": [],
-                        "status": result
-                    }
+                    file.save(file_path)
+
+                    src_third_user_id = user_id
+                    src_category_id = src_img[0]['category_id']
+                    src_file_extension = src_img[0]['file_extension']
+                    new_third_user_id = user_id
+                    new_category_id = category_id
+                    try:
+                        if 0 == mdfs_edit(MDFS_EDIT_URL, MDFS_API_KEY, src_third_user_id, src_category_id, src_title,
+                                          src_file_extension,
+                                          new_third_user_id, new_category_id, new_title, new_file_extension,
+                                          upload_file_name=filename, upload_file_path=file_path):
+                            return {"data": [], "status": -1}
+                    except:
+                        print('文件转存失败')
+                        return {"data": [], "status": -1}
+                    finally:
+                        os.remove(file_path)
+
+                    result = photo.update_photo(user_id, src_title, new_title, category_name)
+                    return {"data": [], "status": result}
                 else:
-                    return {
-                        "data": [],
-                        "status": -1
-                    }
+                    return {"data": [], "status": -1}
         else:
             return redirect(url_for("user_bp.login"))
